best-travel/src/test/java/hu/sinap86/codewars/SumOfKTest.java
package hu.sinap86.codewars;

import static org.junit.Assert.assertEquals;

import org.junit.Test;

import java.util.ArrayList;
import java.util.Arrays;
import java.util.List;

/**
 * @author Sinka László
 */
public class SumOfKTest {

    @Test
    public void BasicTests0() {
        List<Integer> ts = new ArrayList<>(Arrays.asList(50, 55, 57, 58, 60));
        int n = SumOfK.chooseBestSum(174, 3, ts);
        assertEquals(173, n);
    }

    @Test
    public void BasicTests1() {
        System.out.println("****** Basic Tests small numbers******");
        List<Integer> ts = new ArrayList<>(Arrays.asList(50, 55, 56, 57, 58));
        int n = SumOfK.chooseBestSum(163, 3, ts);
        assertEquals(163, n);
        ts = new ArrayList<>(Arrays.asList(50));
        Integer m = SumOfK.chooseBestSum(163, 3, ts);
        assertEquals(null, m);
        ts = new ArrayList<>(Arrays.asList(91, 74, 73, 85, 73, 81, 87));
        n = SumOfK.chooseBestSum(230, 3, ts);
        assertEquals(228, n);
    }
}
